Initial work to turn into a multi-page app
